requirements.txt
fastapi
uvicorn
tensorflow
pillow
pydantic
